rewrite snippet.options
