+++ website_geoengine_store_locator/static/src/js/snippets/s_openstreetmap/snippet.options.js
@@ -1,45 +1,45 @@
-odoo.define("website_snippet_OpenStreetMap.snippet_options", function (require) {
-    "use strict";
+/** @odoo-module **/
 
-    const options = require("web_editor.snippets.options");
-
-    //const publicWidget = require("web.public.widget");
-
-    console.log(options)
+import core from 'web.core';
 
 
-    options.registry.OpenStreetMap = options.Class.extend({
+//import publicWidget from "web.public.widget";
+import dynamicSnippetOptions from 'website.s_dynamic_snippet_options';
 
-        onBuilt() {
-            console.log("onBuilt")
-            this._super.apply(this, arguments);
-            console.log(this.$target[0].id)
-            this.$target[0].id = this.generateUniqueId();
-            console.log(this.$target[0].dataset)
-            //TODO FIND A WAY TO INITIALIZE THE MAP
-            //var map = new OpenLayerMap(this.$target[0].id);
-        },
+import OLM from 'website_geoengine_store_locator.openlayer_map';
 
-        init: function () {
-            this._super.apply(this, arguments);
+const _t = core._t;
 
-        },
+const OpenLayerMap = dynamicSnippetOptions.extend({
 
-        async selectDataAttribute(previewMode, widgetValue, params) {
-            await this._super(...arguments);
-            console.log("Options modified: "+params.attributeName)
-            console.log(params)
-            if (['mapAddress', 'mapType', 'mapZoom'].includes(params.attributeName)) {
-                console.log("Change in map options")
-            }
-        },
-        generateUniqueId: function () {
-            return 'snippet-' + Date.now() + '-' + Math.random().toString(36).substr(2, 9);
-        },
-        saveSnippet: function (previewMode, widgetValue, params) {
-            console.log("save")
-            return this._super(...arguments);
+
+    onBuilt: function () {
+        console.log(OLM)
+        console.log("onBuilt")
+        this._super.apply(this, arguments);
+    },
+
+    saveSnippet: function (previewMode, widgetValue, params) {
+        console.log("save")
+        return this._super(...arguments);
+    },
+
+    selectDataAttribute: async function (previewMode, widgetValue, params) {
+        //await this._super(...arguments);
+        console.log("Options modified: "+params.attributeName)
+        console.log(params)
+        if (['mapAddress', 'mapType', 'mapZoom'].includes(params.attributeName)) {
+            console.log("Change in map options")
         }
+    },
 
-    });
+    cleanForSave: function () {
+        this._super.apply(this, arguments);
+        this.$target.empty();
+    },
+
 });
+
+export default {
+    OpenStreetMap: OpenLayerMap
+};
